pyproject.toml
[project]
name = "foodbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["aiogram",
"asyncpg",
"sqlalchemy[asyncio]",
"python-dotenv",
"aiohttp-socks"
]
